Move Environments tests out of FeaturePreviews tests
